Crate std_unicode

Deprecated since 1.27.0
: moved into libcore

🔬 This is a nightly-only experimental API. (unicode #27783)

The Unicode Library

Unicode-intensive functions for char and str types.

This crate provides a collection of Unicode-related functionality, including decompositions, conversions, etc., and provides traits implementing these functions for the char and str types.

The functionality included here is only that which is necessary to provide for basic string-related manipulations. This crate does not (yet) aim to provide a full set of Unicode tables.

Modules

derived_property 	 [
Experimental
]

property 	 [
Experimental
]